DoNotMerge [AMDGPU] True16 MC Support
This patch supports assembly and disassembly of True16 instructions,
namely adding .l or .h to numbered VGPRs in order to reference their lo
or hi 16 bits.

CodeGen is not functional and contains many errors. Making it functional
is future work. This patch should be used for assembly and disassembly
only.

Also in this patch are changes which may or may not be necessary for MC, but have
been developed for codegen alongside the MC.